Add support for ip rules uidrange option added in Linux 4.10

+++ lib/timer.c
@@ -82,7 +82,7 @@ timer_add(timeval_t a, timeval_t b)
 	ret.tv_usec = a.tv_usec + b.tv_usec;
 	ret.tv_sec = a.tv_sec + b.tv_sec;
 
-	if (ret.tv_usec >= TIMER_HZ) {
+	if (ret.tv_usec >= (int)TIMER_HZ) {
 		ret.tv_sec++;
 		ret.tv_usec -= TIMER_HZ;
 	}
@@ -108,7 +108,7 @@ timer_add_long(timeval_t a, unsigned long b)
 	ret.tv_usec = a.tv_usec + (int)(b % TIMER_HZ);
 	ret.tv_sec = a.tv_sec + (int)(b / TIMER_HZ);
 
-	if (ret.tv_usec >= TIMER_HZ) {
+	if (ret.tv_usec >= (int)TIMER_HZ) {
 		ret.tv_sec++;
 		ret.tv_usec -= TIMER_HZ;
 	}
